move auto-advance to shipping quotes to shipping index
